requirements.txt
requests~=2.32.3
matplotlib~=3.9.2
networkx~=3.3
python-dotenv~=1.0.1
tqdm~=4.66.5
scipy~=1.14.1